fix: When we delete all characters entered before the delimiter and enter new characters, the lookup no longer works.
TinyMCE adds a new span with id "_mce_caret" which wraps the span of the search text :
<span id="_mce_caret" data-mce-bogus="1" data-mce-type="format-caret">
  <span id="autocomplete-searchtext"><span class="dummy">﻿</span></span>
</span>

When we use the backspace key, this key is listened by TinyMCE to check if there are still characters in the current node.

If there is no longer a character, TinyMCE creates a span representing the caret to correctly position the cursor in the editor. Unfortunately, it creates the span with the id "_mce_caret" that wraps the span with the id "autocomplete-searchtext".
And as soon as we enter text again, it is added in the span with the id "autocomplete-delimiter", and the search does not work anymore.

The workaround is to add a second Zero Width No-Break Space (ZWNBSP) to prevent TinyMCE from adding the span caret.

Also:
- add a constant to declare the ZWNBSP character.
- Use editor.selection.setCursorLocation to move the cursor in the span where we enter the search text
- Remove '.replace('\ufeff', '')' because it's not longer present after calling of text() function